split api in api-common and api-public
