+++ .github/workflows/deploy.yml
@@ -34,7 +34,8 @@ jobs:
           src: "build/dokka/htmlMultiModule"
           modules: |
             core
-            api
+            api-common
+            api-public
             sync
             github
           dest: "website/docs"
