+++ api-common/build.gradle.kts
@@ -1,0 +1,10 @@
+val graphglueVersion: String by project
+
+plugins {
+    kotlin("plugin.spring")
+}
+
+dependencies {
+    api(project(path = ":core"))
+    api("io.github.graphglue", "graphglue", graphglueVersion)
+}
