+++ api-common/src/main/kotlin/gropius/graphql/AutoPayloadType.kt
