+++ api-common/src/main/kotlin/gropius/graphql/DefaultSchemaGeneratorHooks.kt
