+++ api-public/src/main/kotlin/gropius/schema/mutation/ArchitectureMutations.kt
